Move common StepsTimeline to steps_modal.
